Smaller batches
- Extract TransformSet to full class
- Move planning code into BatchingTransformManager
- Reduce exposure BatchingTransformManager of internals
- Comment out WaitGroup log :ioa:
- Move ceilingDiv to MoreMath
- Use dynamic chunk size for TransformCalls

+++ src/main/java/com/jozufozu/flywheel/backend/engine/batching/DrawBuffer.java
@@ -75,7 +75,7 @@ public class DrawBuffer {
 		}
 
 		ReusableVertexList vertexList = provider.createVertexList();
-		vertexList.ptr(memory.ptr() + startVertex * stride);
+		vertexList.ptr(memory.ptr() + (long) startVertex * stride);
 		vertexList.vertexCount(vertexCount);
 		return vertexList;
 	}
